Removed test files and added ease of use to run.py
